Add status route

+++ rest-api/api/controllers/feed.js
@@ -7,6 +7,7 @@ const { validationResult } = require('express-validator');
 
 // Models
 const Post = require('../models/post');
+const User = require('../models/user');
 
 // GET
 exports.getPosts = (req, res, next) => {
@@ -14,7 +15,7 @@ exports.getPosts = (req, res, next) => {
     console.log('page query', page);
     const perPage = 2;
     let totalItems;
-    Post.countDocuments()
+    Post.find().countDocuments()
         .then(numberOfDocs => {
             totalItems = numberOfDocs;
             return Post.find()
@@ -81,21 +82,32 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path;
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
 
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: {
-            name: 'Diogo'
-        }
+        creator: req.userId
     });
 
     post.save()
         .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
+        .then(result => {
             res.status(201).json({
                 message: 'The post was created successfully',
-                post: result
+                post: post,
+                creator: {
+                    _id: creator._id,
+                    name: creator.name
+                }
             });
         })
         .catch(e => {
@@ -135,6 +147,12 @@ exports.updatePost = (req, res, next) => {
                 throw error;
             }
 
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
+
             // Delete old image
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
@@ -166,12 +184,26 @@ exports.deletePost = (req, res, next) => {
     Post.findById(id)
         .then(post => {
             if (!post) {
-                const error = new Error('Could not found post');
+                const error = new Error('Could not find post');
                 error.statusCode = 404;
                 throw error;
             }
+
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
+
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(id);
+        })
+        .then(() => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.post.pull(id);
+            return user.save();
         })
         .then(() => {
             res.status(200).json({
